+++ src/components/Accordion/Accordion.test.tsx
@@ -388,7 +388,7 @@ describe("Accordion", () => {
     const withError = renderer
       .create(
         <Accordion
-          accordionData={[{ ...accordionData[0], accordionType: "error" }]}
+          accordionData={[{ ...accordionData[0], variant: "error" }]}
           id="accordian"
           isDefaultOpen
         />
@@ -397,7 +397,7 @@ describe("Accordion", () => {
     const withWarning = renderer
       .create(
         <Accordion
-          accordionData={[{ ...accordionData[0], accordionType: "warning" }]}
+          accordionData={[{ ...accordionData[0], variant: "warning" }]}
           id="accordian"
           isDefaultOpen
         />