+++ src/components/Breadcrumbs/Breadcrumbs.test.tsx
@@ -151,7 +151,7 @@ describe("Breadcrumbs Snapshot", () => {
       .create(
         <Breadcrumbs
           breadcrumbsData={breadcrumbsData}
-          breadcrumbsType="booksAndMore"
+          variant="booksAndMore"
           id="breadcrumbs-test"
         />
       )
@@ -160,7 +160,7 @@ describe("Breadcrumbs Snapshot", () => {
       .create(
         <Breadcrumbs
           breadcrumbsData={breadcrumbsData}
-          breadcrumbsType="blogs"
+          variant="blogs"
           id="breadcrumbs-test"
         />
       )
@@ -169,7 +169,7 @@ describe("Breadcrumbs Snapshot", () => {
       .create(
         <Breadcrumbs
           breadcrumbsData={breadcrumbsData}
-          breadcrumbsType="locations"
+          variant="locations"
           id="breadcrumbs-test"
         />
       )
@@ -178,7 +178,7 @@ describe("Breadcrumbs Snapshot", () => {
       .create(
         <Breadcrumbs
           breadcrumbsData={breadcrumbsData}
-          breadcrumbsType="education"
+          variant="education"
           id="breadcrumbs-test"
         />
       )