+++ src/components/SearchBar/SearchBar.tsx
@@ -151,7 +151,7 @@ export const SearchBar: ChakraComponent<
         labelText={selectProps?.labelText}
         name={selectProps?.name}
         onChange={selectProps?.onChange}
-        selectType="searchbar"
+        variant="searchbar"
         value={selectProps?.value}
         __css={styles.select}
         data-select
@@ -181,7 +181,7 @@ export const SearchBar: ChakraComponent<
         onChange={textInputProps?.onChange}
         pattern={textInputProps?.pattern}
         placeholder={textInputPlaceholder}
-        textInputType={selectElem ? "searchBarSelect" : "searchBar"}
+        textInputVariant={selectElem ? "searchBarSelect" : "searchBar"}
         type="text"
         value={textInputProps?.value}
         {...stateProps}
@@ -191,7 +191,7 @@ export const SearchBar: ChakraComponent<
     const buttonElem = (
       <Button
         className="searchButton"
-        buttonType={buttonType}
+        variant={buttonType}
         id={`searchbar-button-${id}`}
         isDisabled={isDisabled}
         onClick={buttonOnClick}