+++ src/components/StatusBadge/StatusBadge.tsx
@@ -13,13 +13,12 @@ export const statusBadgeFontSizeArray = [
   "desktop.body.body2",
   "desktop.caption",
 ] as const;
-
-export type StatusBadgeTypes = typeof messageVariantsArray[number];
+export type StatusBadgeVariants = typeof messageVariantsArray[number];
 export interface StatusBadgeProps extends BoxProps {
   /** ID that other components can cross reference for accessibility purposes */
   id?: string;
   /** Semantic type of the status badge. */
-  type?: StatusBadgeTypes;
+  variant?: StatusBadgeVariants;
 }
 
 /**
@@ -36,10 +35,10 @@ export const StatusBadge: ChakraComponent<
 > = chakra(
   forwardRef<HTMLDivElement, React.PropsWithChildren<StatusBadgeProps>>(
     (props, ref?) => {
-      const { children, id, type, ...rest } = props;
+      const { children, id, variant, ...rest } = props;
       const styles = useStyleConfig("StatusBadge", {
         labelFontSize: rest["fontSize"],
-        variant: type || "neutral",
+        variant: variant || "neutral",
       });
 
       if (!children) {