+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Changes the names of theme objects that conflict with Chakra's base names from `Custom` to `Reservoir` for consistency. This impacts theme objects for `Breadcrumbs`, `Button`, `Select`, `Slider`, and `Table`.
 - Replaces positional function args with object references in utility functions/components when there are 3 or more arguments.
 - Updates `Table`'s `tableTextSize` prop to accept `caption`.
+- Renames `Accordion`, `Banner`, `Breadcrumbs`, `Button`, `Hero`, `Icon`, `Link`, `List`, `Modal`, `Notification`, `Select`, `StatusBadge`, `TagSet`, and `TextInput` props using "type" to use "variant".
 
 ### Removals
 
@@ -49,6 +50,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Removes `useNYPLBreakpoints` from final component, `SearchBar`, and removes `mediaMatchMock` from test setup since it is no longer necessary.
 - Removes explicit `className` and `children` props in favor of expanded prop type definitions in all components.
 - Removes the use of the `window`'s `scrollIntoView` dependency in the `Tabs` component.
+- Removes the `tableTextSize` prop from the `Table` component in favor of using the native `fontSize` CSS prop.
 
 ## Prerelease